backend/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Local mode: Vector memory (ChromaDB)
chromadb>=0.4.0

# Optional: Cloud mode PostgreSQL support
# Uncomment for cloud deployment
# psycopg2-binary>=2.9.0
